// packages/backend/convex/_generated/api.d.ts
declare const fullApi: ApiFromModules<{
  agent: typeof agent;
  chat: typeof chat;
  documents: typeof documents;
  healthCheck: typeof healthCheck;
  indexer: typeof indexer;
  migrations: typeof migrations;
  ocr: typeof ocr;
  privateData: typeof privateData;
  search: typeof search;
}>;
// ...
export declare const internal: FilterApi<
  typeof fullApi,
  FunctionReference<any, "internal">
>;